requirements.txt
boto3==1.20.48
botocore==1.23.48
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
jmespath==0.10.0
lxml==4.7.1
Pillow==9.0.1
pip-autoremove==0.10.0
python-dateutil==2.8.2
python-pptx==0.6.21
requests==2.27.1
rt==2.2.1
s3transfer==0.5.1
six==1.16.0
urllib3==1.26.8
XlsxWriter==3.0.2
